Update line follower
The Line Follower Robot is an autonomous robotic system designed to follow a predefined path, usually a black line on a white surface or vice versa. It achieves this by utilizing infrared (IR) sensors to detect line patterns and an L298N motor driver to control the movement of two DC motors.